# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a real-time social discovery engine.",
        "title": "Radar Momentum API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/plans": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plans"
                ],
                "summary": "Create a new plan",
                "parameters": [
                    {
                        "description": "Plan data",
                        "name": "plan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_xanderhill_radar_internal_models.Plan"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_xanderhill_radar_internal_models.Plan"
                        }
                    }
                }
            }
        },
        "/plans/{id}/checkin": {
            "post": {
                "description": "Increments the check-in count for a specific plan",
                "tags": [
                    "plans"
                ],
                "summary": "Check in to a plan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Plan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/plans/{id}/save": {
            "post": {
                "description": "Increments the save count for a specific plan",
                "tags": [
                    "plans"
                ],
                "summary": "Save a plan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Plan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/radar": {
            "get": {
                "description": "Fetches plans within a radius and sorts by momentum score",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plans"
                ],
                "summary": "Get nearby trending plans",
                "parameters": [
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "Latitude",
                        "name": "lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "Longitude",
                        "name": "lng",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter by category (e.g. coffee, art)",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Includes distance_meters and momentum_score",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_xanderhill_radar_internal_models.Plan"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_xanderhill_radar_internal_models.Plan": {
            "type": "object",
            "properties": {
                "base_score": {
                    "description": "Analytics signals for the Momentum Engine",
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "check_ins": {
                    "description": "Real-time presence",
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "distance_meters": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "last_checkin_at": {
                    "type": "string"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "momentum_score": {
                    "description": "Enriched Fields",
                    "type": "number"
                },
                "saves": {
                    "description": "Intent signal",
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}